# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PostPeer unified social media API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalPosts: number;
  publishedPosts: number;
  scheduledPosts: number;
  failedPosts: number;
  connectedPlatforms: number;
  totalDeliveries: number;
  creditsUsed: number;
  creditsTotal: number;
}

export type PostStatus = typeof PostStatus[keyof typeof PostStatus];


export const PostStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  failed: 'failed',
} as const;

export interface Post {
  id: number;
  userId: string;
  content: string;
  platforms: string[];
  status: PostStatus;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  mediaUrls: string[];
  createdAt: string;
  updatedAt: string;
}

export interface PostList {
  posts: Post[];
  total: number;
}

export interface PostInput {
  /** @minLength 1 */
  content: string;
  /** @minItems 1 */
  platforms: string[];
  /** @nullable */
  scheduledAt?: string | null;
  mediaUrls?: string[];
}

export interface Platform {
  id: string;
  name: string;
  icon: string;
  color: string;
  maxLength: number;
  supportsMedia: boolean;
}

export interface PlatformList {
  platforms: Platform[];
}

export type ConnectedPlatformStatus = typeof ConnectedPlatformStatus[keyof typeof ConnectedPlatformStatus];


export const ConnectedPlatformStatus = {
  connected: 'connected',
  disconnected: 'disconnected',
} as const;

export interface ConnectedPlatform {
  id: number;
  platform: string;
  accountName: string;
  accountHandle: string;
  status: ConnectedPlatformStatus;
  connectedAt: string;
}

export interface ConnectedPlatformList {
  platforms: ConnectedPlatform[];
}

export interface PlatformConnectInput {
  platform: string;
  accountName: string;
  accountHandle: string;
}

export interface ApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
  active: boolean;
}

export interface ApiKeyList {
  keys: ApiKey[];
}

export interface ApiKeyInput {
  /** @minLength 1 */
  name: string;
}

export interface ApiKeyCreated {
  id: number;
  name: string;
  keyPrefix: string;
  fullKey: string;
  createdAt: string;
  active: boolean;
}

export type UserProfilePlan = typeof UserProfilePlan[keyof typeof UserProfilePlan];


export const UserProfilePlan = {
  free: 'free',
  starter: 'starter',
  standard: 'standard',
  pro: 'pro',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  plan: UserProfilePlan;
  createdAt: string;
}

export type ListPostsParams = {
status?: ListPostsStatus;
platform?: string;
limit?: number;
offset?: number;
};

export type ListPostsStatus = typeof ListPostsStatus[keyof typeof ListPostsStatus];


export const ListPostsStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  failed: 'failed',
} as const;