# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PostPeer unified social media API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
  active: boolean;
}