# lib/api-zod/src/generated/types/apiKeyInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PostPeer unified social media API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyInput {
  /** @minLength 1 */
  name: string;
}