# lib/api-zod/src/generated/types/apiKeyList.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PostPeer unified social media API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKey } from './apiKey';

export interface ApiKeyList {
  keys: ApiKey[];
}