# lib/api-zod/src/generated/types/connectedPlatform.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PostPeer unified social media API
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectedPlatformStatus } from './connectedPlatformStatus';

export interface ConnectedPlatform {
  id: number;
  platform: string;
  accountName: string;
  accountHandle: string;
  status: ConnectedPlatformStatus;
  connectedAt: string;
}