# lib/api-zod/src/generated/types/connectedPlatformList.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PostPeer unified social media API
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectedPlatform } from './connectedPlatform';

export interface ConnectedPlatformList {
  platforms: ConnectedPlatform[];
}