# lib/api-zod/src/generated/types/platformList.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * PostPeer unified social media API
 * OpenAPI spec version: 0.1.0
 */
import type { Platform } from './platform';

export interface PlatformList {
  platforms: Platform[];
}